# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Student Learning Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  student: 'student',
  client: 'client',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  student: 'student',
  client: 'client',
} as const;

export interface UserInput {
  email: string;
  name: string;
  role: UserInputRole;
  password: string;
}

export interface AdminStats {
  totalUsers: number;
  totalStudents: number;
  totalAdmins: number;
  totalClients: number;
  recentUsers: User[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export type QuizDifficulty = typeof QuizDifficulty[keyof typeof QuizDifficulty];


export const QuizDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Question {
  id: number;
  text: string;
  options: string[];
}

export interface Quiz {
  id: number;
  title: string;
  subject: string;
  questionCount: number;
  difficulty: QuizDifficulty;
  questions?: Question[];
}

export interface QuizSubmission {
  answers: number[];
}

export interface QuizResult {
  quizId: number;
  score: number;
  totalQuestions: number;
  correct: number;
  passed: boolean;
  completedAt: string;
}

export interface ProgressOverview {
  totalQuizzes: number;
  averageScore: number;
  topSubject: string;
  streak: number;
  completionRate: number;
}

export interface QuizReport {
  id: number;
  quizTitle: string;
  subject: string;
  score: number;
  passed: boolean;
  completedAt: string;
}

export interface PerformancePoint {
  date: string;
  score: number;
  subject: string;
}