# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Student Learning Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AdminStats {
  totalUsers: number;
  totalStudents: number;
  totalAdmins: number;
  totalClients: number;
  recentUsers: User[];
}