# lib/api-zod/src/generated/types/progressOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Student Learning Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressOverview {
  totalQuizzes: number;
  averageScore: number;
  topSubject: string;
  streak: number;
  completionRate: number;
}