# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Student Learning Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';
import type { QuizDifficulty } from './quizDifficulty';

export interface Quiz {
  id: number;
  title: string;
  subject: string;
  questionCount: number;
  difficulty: QuizDifficulty;
  questions?: Question[];
}