# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Student Learning Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UserInputRole } from './userInputRole';

export interface UserInput {
  email: string;
  name: string;
  role: UserInputRole;
  password: string;
}